telegram-notion-scheduler/requirements.txt
python-telegram-bot==20.7
notion-client==2.2.1
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
psutil==5.9.6
